#211103_백정연_회원정보 수정 메소드 수정 commit date : 211103 writer : 백정연 branch : 210903_jeong_optimizing comment - 회원 정보 수정 후 마이페이지 이동시 글목록 페이징 번호 오류 수정

+++ src/main/java/com/finalprj/doldolseo/controller/mypage/PlannerApiController.java
@@ -75,6 +75,10 @@ public class PlannerApiController {
     // 플래너 삭제
     @DeleteMapping("/users/{userid}/planners/{plannerid}")
     public ResponseEntity<String> deletePlanner(@PathVariable("plannerid") long plannerId, @PathVariable("userid") String userId) {
+        PlannerDTO dto = plannerService.selectPlanner(plannerId);
+        if(dto == null){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("해당 플래너가 존재하지 않습니다, 다시 확인해주세요.");
+        }
         planService.deletePlans(plannerId);
         plannerService.deletePlanner(plannerId);
